Renamed file useTokensLazy to useTokenLazy
